Move CodeWrapper to CodeBlock directory
